+++ Fuji/Source/Drivers/Linux/MFSound_PulseAudio.cpp
@@ -305,7 +305,7 @@ MF_API void MFSound_SetMasterVolume(float volume)
 */
 }
 
-uint32 MFSound_GetPlayCursorInternal(MFVoice *pVoice, uint32 *pWriteCursor)
+size_t MFSound_GetPlayCursorInternal(MFVoice *pVoice, size_t *pWriteCursor)
 {
 	MFCALLSTACK;
