+++ Fuji/Source/Drivers/Sound/MFSound_OpenAL.cpp
@@ -590,7 +590,7 @@ MF_API void MFSound_SetMasterVolume(float volume)
 	alListenerf(AL_GAIN, volume);
 }
 
-uint32 MFSound_GetPlayCursorInternal(MFVoice *pVoice, uint32 *pWriteCursor)
+size_t MFSound_GetPlayCursorInternal(MFVoice *pVoice, size_t *pWriteCursor)
 {
 	MFCALLSTACK;
